requirements.txt
# ARTIS - Autonomous Red Teaming Integrated System
# Phase 1 MVP Dependencies

# CLI Framework
cmd2>=2.4.0
colorama>=0.4.6
tabulate>=0.9.0
click>=8.1.0
tqdm>=4.66.0

# Message Bus
pika>=1.3.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Logging
python-json-logger>=2.0.0

# STIX 2.x
stix2>=3.0.0

# Tool Integration
python-libnmap>=0.7.3
requests>=2.31.0
python-owasp-zap-v2.4>=0.0.21
pymetasploit3>=1.0.3

# Utilities
xmltodict>=0.13.0
validators>=0.22.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
